Rename src/testinfra -> src/testutils
This solves the problem of a duplicate wheel name
