+++ src/python/pants/testutils/engine/base_engine_test.py
@@ -3,7 +3,7 @@
 
 from pants.goal.goal import Goal
 from pants.goal.task_registrar import TaskRegistrar
-from pants.testinfra.test_base import TestBase
+from pants.testutils.test_base import TestBase
 
 
 class EngineTestBase(TestBase):
