+++ src/python/pants/testutils/jvm/jar_task_test_base.py
@@ -1,7 +1,7 @@
 # Copyright 2019 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.testinfra.jvm.nailgun_task_test_base import NailgunTaskTestBase
+from pants.testutils.jvm.nailgun_task_test_base import NailgunTaskTestBase
 
 
 class JarTaskTestBase(NailgunTaskTestBase):
