+++ src/python/pants/testutils/jvm/jvm_task_test_base.py
@@ -5,8 +5,8 @@ import os
 
 from pants.backend.jvm.subsystems.resolve_subsystem import JvmResolveSubsystem
 from pants.backend.jvm.tasks.classpath_products import ClasspathProducts
-from pants.testinfra.subsystem.util import init_subsystem
-from pants.testinfra.task_test_base import TaskTestBase
+from pants.testutils.subsystem.util import init_subsystem
+from pants.testutils.task_test_base import TaskTestBase
 from pants.util.dirutil import safe_file_dump, safe_mkdir, safe_mkdtemp
