+++ tests/python/pants_test/base/context_utils.py
@@ -1,7 +1,7 @@
 # Copyright 2014 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.testinfra.base.context_utils import TestContext as TestContext
-from pants.testinfra.base.context_utils import (
+from pants.testutils.base.context_utils import TestContext as TestContext  # noqa
+from pants.testutils.base.context_utils import (
   create_context_from_options as create_context_from_options,
-)
+)  # noqa
