+++ tests/python/pants_test/console_rule_test_base.py
@@ -1,4 +1,4 @@
 # Copyright 2018 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.testinfra.console_rule_test_base import ConsoleRuleTestBase as ConsoleRuleTestBase
+from pants.testutils.console_rule_test_base import ConsoleRuleTestBase as ConsoleRuleTestBase  # noqa
