+++ tests/python/pants_test/engine/base_engine_test.py
@@ -1,4 +1,4 @@
 # Copyright 2014 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.testinfra.engine.base_engine_test import EngineTestBase as EngineTestBase
+from pants.testutils.engine.base_engine_test import EngineTestBase as EngineTestBase  # noqa
