+++ tests/python/pants_test/engine/util.py
@@ -1,13 +1,13 @@
 # Copyright 2016 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.testinfra.engine.util import TARGET_TABLE as TARGET_TABLE
-from pants.testinfra.engine.util import MockConsole as MockConsole
-from pants.testinfra.engine.util import Target as Target
-from pants.testinfra.engine.util import assert_equal_with_printing as assert_equal_with_printing
-from pants.testinfra.engine.util import create_scheduler as create_scheduler
-from pants.testinfra.engine.util import init_native as init_native
-from pants.testinfra.engine.util import (
+from pants.testutils.engine.util import TARGET_TABLE as TARGET_TABLE  # noqa
+from pants.testutils.engine.util import MockConsole as MockConsole  # noqa
+from pants.testutils.engine.util import Target as Target  # noqa
+from pants.testutils.engine.util import assert_equal_with_printing as assert_equal_with_printing  # noqa
+from pants.testutils.engine.util import create_scheduler as create_scheduler  # noqa
+from pants.testutils.engine.util import init_native as init_native  # noqa
+from pants.testutils.engine.util import (
   remove_locations_from_traceback as remove_locations_from_traceback,
-)
-from pants.testinfra.engine.util import run_rule as run_rule
+)  # noqa
+from pants.testutils.engine.util import run_rule as run_rule  # noqa
