+++ tests/python/pants_test/jvm/BUILD
@@ -23,7 +23,7 @@ python_library(
   name='jvm_tool_task_test_base',
   sources=['jvm_tool_task_test_base.py'],
   dependencies=[
-    'src/python/pants/testinfra/jvm:jvm_tool_task_test_base',
+    'src/python/pants/testutils/jvm:jvm_tool_task_test_base',
   ]
 )
 
@@ -31,7 +31,7 @@ python_library(
   name='nailgun_task_test_base',
   sources=['nailgun_task_test_base.py'],
   dependencies=[
-    'src/python/pants/testinfra/jvm:nailgun_task_test_base',
+    'src/python/pants/testutils/jvm:nailgun_task_test_base',
   ]
 )
 
@@ -39,7 +39,7 @@ python_library(
   name='jar_task_test_base',
   sources=['jar_task_test_base.py'],
   dependencies=[
-    'src/python/pants/testinfra/jvm:jar_task_test_base',
+    'src/python/pants/testutils/jvm:jar_task_test_base',
   ]
 )
 
@@ -47,6 +47,6 @@ python_library(
   name='jvm_task_test_base',
   sources=['jvm_task_test_base.py'],
   dependencies=[
-    'src/python/pants/testinfra/jvm:jvm_task_test_base',
+    'src/python/pants/testutils/jvm:jvm_task_test_base',
   ]
 )
