+++ tests/python/pants_test/jvm/jar_task_test_base.py
@@ -1,4 +1,4 @@
 # Copyright 2014 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.testinfra.jvm.jar_task_test_base import JarTaskTestBase as JarTaskTestBase
+from pants.testutils.jvm.jar_task_test_base import JarTaskTestBase as JarTaskTestBase  # noqa
