+++ tests/python/pants_test/jvm/jvm_task_test_base.py
@@ -1,4 +1,4 @@
 # Copyright 2015 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.testinfra.jvm.jvm_task_test_base import JvmTaskTestBase as JvmTaskTestBase
+from pants.testutils.jvm.jvm_task_test_base import JvmTaskTestBase as JvmTaskTestBase  # noqa
