+++ tests/python/pants_test/jvm/jvm_tool_task_test_base.py
@@ -1,4 +1,4 @@
 # Copyright 2014 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.testinfra.jvm.jvm_tool_task_test_base import JvmToolTaskTestBase as JvmToolTaskTestBase
+from pants.testutils.jvm.jvm_tool_task_test_base import JvmToolTaskTestBase as JvmToolTaskTestBase  # noqa
