+++ tests/python/pants_test/jvm/nailgun_task_test_base.py
@@ -1,4 +1,4 @@
 # Copyright 2014 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.testinfra.jvm.nailgun_task_test_base import NailgunTaskTestBase as NailgunTaskTestBase
+from pants.testutils.jvm.nailgun_task_test_base import NailgunTaskTestBase as NailgunTaskTestBase  # noqa
