+++ tests/python/pants_test/option/util/fakes.py
@@ -1,11 +1,11 @@
 # Copyright 2015 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.testinfra.option.fakes import _FakeOptionValues as _FakeOptionValues
-from pants.testinfra.option.fakes import (
+from pants.testutils.option.fakes import _FakeOptionValues as _FakeOptionValues  # noqa
+from pants.testutils.option.fakes import (
   _options_registration_function as _options_registration_function,
-)
-from pants.testinfra.option.fakes import create_options as create_options
-from pants.testinfra.option.fakes import (
+)  # noqa
+from pants.testutils.option.fakes import create_options as create_options  # noqa
+from pants.testutils.option.fakes import (
   create_options_for_optionables as create_options_for_optionables,
-)
+)  # noqa
