+++ tests/python/pants_test/pants_run_integration_test.py
@@ -1,13 +1,13 @@
 # Copyright 2014 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.testinfra.pants_run_integration_test import PantsJoinHandle as PantsJoinHandle
-from pants.testinfra.pants_run_integration_test import PantsResult as PantsResult
-from pants.testinfra.pants_run_integration_test import (
+from pants.testutils.pants_run_integration_test import PantsJoinHandle as PantsJoinHandle  # noqa
+from pants.testutils.pants_run_integration_test import PantsResult as PantsResult  # noqa
+from pants.testutils.pants_run_integration_test import (
   PantsRunIntegrationTest as PantsRunIntegrationTest,
-)
-from pants.testinfra.pants_run_integration_test import ensure_cached as ensure_cached
-from pants.testinfra.pants_run_integration_test import ensure_daemon as ensure_daemon
-from pants.testinfra.pants_run_integration_test import ensure_resolver as ensure_resolver
-from pants.testinfra.pants_run_integration_test import read_pantsd_log as read_pantsd_log
-from pants.testinfra.pants_run_integration_test import render_logs as render_logs
+)  # noqa
+from pants.testutils.pants_run_integration_test import ensure_cached as ensure_cached  # noqa
+from pants.testutils.pants_run_integration_test import ensure_daemon as ensure_daemon  # noqa
+from pants.testutils.pants_run_integration_test import ensure_resolver as ensure_resolver  # noqa
+from pants.testutils.pants_run_integration_test import read_pantsd_log as read_pantsd_log  # noqa
+from pants.testutils.pants_run_integration_test import render_logs as render_logs  # noqa
