+++ tests/python/pants_test/subsystem/subsystem_util.py
@@ -1,6 +1,6 @@
 # Copyright 2015 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.testinfra.subsystem.util import global_subsystem_instance as global_subsystem_instance
-from pants.testinfra.subsystem.util import init_subsystem as init_subsystem
-from pants.testinfra.subsystem.util import init_subsystems as init_subsystems
+from pants.testutils.subsystem.util import global_subsystem_instance as global_subsystem_instance  # noqa
+from pants.testutils.subsystem.util import init_subsystem as init_subsystem  # noqa
+from pants.testutils.subsystem.util import init_subsystems as init_subsystems  # noqa
