+++ tests/python/pants_test/task_test_base.py
@@ -1,8 +1,8 @@
 # Copyright 2014 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.testinfra.task_test_base import ConsoleTaskTestBase as ConsoleTaskTestBase
-from pants.testinfra.task_test_base import DeclarativeTaskTestMixin as DeclarativeTaskTestMixin
-from pants.testinfra.task_test_base import TaskTestBase as TaskTestBase
-from pants.testinfra.task_test_base import ensure_cached as ensure_cached
-from pants.testinfra.task_test_base import is_exe as is_exe
+from pants.testutils.task_test_base import ConsoleTaskTestBase as ConsoleTaskTestBase  # noqa
+from pants.testutils.task_test_base import DeclarativeTaskTestMixin as DeclarativeTaskTestMixin  # noqa
+from pants.testutils.task_test_base import TaskTestBase as TaskTestBase  # noqa
+from pants.testutils.task_test_base import ensure_cached as ensure_cached  # noqa
+from pants.testutils.task_test_base import is_exe as is_exe  # noqa
