+++ tests/python/pants_test/test_base.py
@@ -1,5 +1,5 @@
 # Copyright 2018 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.testinfra.test_base import AbstractTestGenerator as AbstractTestGenerator
-from pants.testinfra.test_base import TestBase as TestBase
+from pants.testutils.test_base import AbstractTestGenerator as AbstractTestGenerator  # noqa
+from pants.testutils.test_base import TestBase as TestBase  # noqa
