+++ tests/python/pants_test/testutils/mock_logger.py
@@ -1,4 +1,4 @@
 # Copyright 2014 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.testinfra.utils.mock_logger import MockLogger as MockLogger
+from pants.testutils.utils.mock_logger import MockLogger as MockLogger  # noqa
