+++ tests/python/pants_test/testutils/process_test_util.py
@@ -1,14 +1,14 @@
 # Copyright 2017 Pants project contributors (see CONTRIBUTORS.md).
 # Licensed under the Apache License, Version 2.0 (see LICENSE).
 
-from pants.testinfra.utils.process_test_util import ProcessStillRunning as ProcessStillRunning
-from pants.testinfra.utils.process_test_util import (
+from pants.testutils.utils.process_test_util import ProcessStillRunning as ProcessStillRunning  # noqa
+from pants.testutils.utils.process_test_util import (
   TrackedProcessesContext as TrackedProcessesContext,
-)
-from pants.testinfra.utils.process_test_util import _make_process_table as _make_process_table
-from pants.testinfra.utils.process_test_util import (
+)  # noqa
+from pants.testutils.utils.process_test_util import _make_process_table as _make_process_table  # noqa
+from pants.testutils.utils.process_test_util import (
   _safe_iter_matching_processes as _safe_iter_matching_processes,
-)
-from pants.testinfra.utils.process_test_util import (
+)  # noqa
+from pants.testutils.utils.process_test_util import (
   no_lingering_process_by_command as no_lingering_process_by_command,
-)
+)  # noqa
